Clean queries folders so only one Rental folder
